Add styles for desired outcomes
Add 8px padding to the header class, and apply a flex value to the
display property to make the header into a flex container, then use
align-items with a value of center to vertically place the flex items in
the header in the center of the parent element/flex container.
Apply flex to the ul to create a flex container, and add a gap of 16px
to separate those flex items from each other, and finally apply a center
value to the align items property to vertically center those flex items.
Create a rule set for our new div element which surrounds the button and
profile image div so that we can align them to the right side of the
header using margin-left: auto; and apply flex to its display property
which makes it into a flex container, again apply a 16px gap between
it's flex items, and align them center on the vertical axis.