Add the correct styles
Instead of what I applied before, the solution applies a justify-content
of space-between, which when we add 2 new divs places the space between
them, and even though we were told to center vertically their solution
did not use align-items: center.
I forgot to eliminate the default margin and padding that the ul element
had.
Again instead of applying just one div for the two elements that will be
positioned on the right of the header they add 2 new divs, one for each
side of the header and within each of these divs they apply flex to
turn each div into a flex container and apply an align-items: center; to
position those flex items vertically centerec, and they apply a 16px gap
between those flex items.