Add styles to adjust new div elements
We'll need to center text on the entire page so add text-align: center;
to the body element, and that will be inherited by its descendant child
elements on the page.
Apply a margin-bottom of 32 px to the title class which vertically
separates the title from the plant elements.
Apply styles to the first div we added which is the container, this will
become a flex container so we use display: flex;, then we use need to
center horizontally the items within that container thus use center to
justify-content, and we need a 52px gap between the plant divs inside
this container so gap: 52px;
Finally we'll need to set a max-width of 200px to the individual plant
elements/divs so that they don't get larger than what we want.