Add a div, and move an element into another
We'll need to add a div so that we can group together the heading, the
close button, the text, and the continue and cancel buttons, then add
styles to it to make it a flex container so we can adjust the layout
of its child elements.
We'll need to separate that div from the icon class div which will be
just the exclamation mark.
I gave up but was close, I see now that I was very off from the solution given.