Add styles to css file
Position header and footer to the top and bottom of the page, as that is
typical design, also flex those elements so we can position their
children elements and place them on the left and right.
Also give header and footer padding.
Color the background of the footer to a gray #eeeeee
Change the divs in the header and footer to be flex containers, and
eliminate the list item markers, and give those flex items gaps of 10px.

Change the anchor elements that are inside list items to be a gray color
text and remove the default color of links.

Change the content class to be a flex container and declare its flex
direction to be a column and then change align-items to center so we can
horizontaly center the flex items of the container class container.
Add padding to the button class elements as that will make them look
better.
Add margin to the elements inside the content class container to
separate them rom each other.
Remove the padding and margin from the ul, aka the default styles.