Add div elements to separate content
Add a div element around the card class div elements, so we can layout
those cards efficiently via flex in that added div.
Add a div around the sidebar and the newly created card wrapping div, so
we can style the middle of our page correctly away from the header and
footer.