Add styles to correctly format the layout
To change the orientation of the elements on the page we need to use a
flex display with a direction of column on the body element.
To appropriately position the header atop the page we'll need to use
align-self to start, and for it to take up the entire width of the page
give it a width of 100%. We'll also need the text content inside of it
to be thick so a font-weight of 800 should do. And to place the text
appropriately we'll need to change the header to a flex container and
use align-items property with a value of center to center it vertically.
We'll also need to change the font size to 32px and give the content a
padding on its left side of 16px to space it away from the headers edge.
The footer will also need to be changed to a flex container so we can
appropriately position its text content in the center vertically and
horizontally via justify-content and align-self with center values. And
to take up the entire width of the page we pass it a value of 100%. Also
we'll need to make sure it stays at the bottom of the page so we'll use
an align-self value of end to accomplish that.
To position the sidebar effectively we'll need to give it a height of
100% which is the height available to it via the content classed element
which sets between the header and footer. We'll give it a width of 300px
and disallow flex to shrink it. And we'll need to give the element a
padding of 16px.
Then we'll need to change the unordered list inside that sidebar so the
list items don't have a marker, and we'll remove the default padding and
margin as well.
The anchor links in that unordered list will have their default
underlines removed, and text color changed to white, and their font size
changed to 24px.
The elements with the class name of card will have a flex grow of 1 set
and a flex basis of 300px also set which acts as a desired starting
width.
The content class element which surrounds the card classed elements will
be a flex container, with flex wrap, and a gap between its flex items of
20px, and it will padding of 48px, and it's height will be as much as
its parent element is offering it.
Finally the middle classed element, which holds the sidebar and content
classed element and sits between the header and footer will be a flex
container and have a flex-grow of 1.