Add solutions to the assignments, etc
I removed a line in their solution file that was throwing me off, it
was just a blank line but none the less it didn't exist in their
starter css file.
I took a while playing with rows, as their solution to the second
assignment didn't have explicit rows defined yet used grid-row on the
sidebar and that confused me.
I also played around with setting a width and height on the container
of the 100vw and 100vh values, the solution didn't use anything close
to that so I'm a little confused because when I didn't state them my
result would get bonkers, but I'll have to play with grid some more.